Fix buffer size bug in getlibcaddr()
Basically, when I read a line using fgets(), I had a buffer that was 850
bytes long, but I was only using 85 bytes of it. 85 bytes apparently
wasn't enough to include the string that says "libc" when reading lines
of /proc/pid/maps. As a result, getlibcaddr() was finding the wrong
address of libc for the target.

Since the libc address was wrong, we were calculating function addresses
that were completely wrong as well. The injector was calling random
addresses inside the target's address space, which predictably caused it
to segfault.

Even more confusingly, this issue was only happening for me on Arch and
not on Ubuntu, so for a while I thought it was some kind of
configuration issue.