arm: use refactored ptrace.c (see details below)
arm's register type is user_regs, while it's user_regs_struct on x86 and
x86_64, so I had to do some defines in order to decide which type to use
at build time. other than that, the factored ptrace code seems to work
fine on all 3 architectures, as far as I can see.